# docs/openapi/chimely.yaml
openapi: 3.0.3
info:
  title: Chimely API
  description: |
    Two planes, one binary:

    * **Management plane** — called by the customer's backend with a Bearer API
      key. Creates notifications and broadcasts, manages subscribers and their
      preferences. API keys are environment-scoped; the environment is implied
      by the key.
    * **Subscriber plane** — called by `@chimely/client` (the `<Inbox />`
      widget) on behalf of one end user. Authenticated with an HMAC subscriber
      hash: `hex(HMAC-SHA256(environment.subscriber_hmac_secret, subscriber_id))`
      computed by the customer's backend. Mandatory in environments with
      `require_subscriber_hash = true` (the production default); optional in
      dev environments so the quickstart works without a backend.

    Subscriber-plane scoping travels in headers (or query parameters where
    headers are impossible, i.e. `EventSource`):

    | Header | Query fallback | Meaning |
    |---|---|---|
    | `X-Chimely-Environment` | `environment` | environment slug |
    | `X-Chimely-Subscriber` | `subscriber_id` | customer-provided subscriber id |
    | `X-Chimely-Subscriber-Hash` | `subscriber_hash` | HMAC hash (when required) |

    **Idempotency.** Every create accepts `idempotency_key` (client-supplied or
    server-generated and echoed). Uniqueness is per environment per resource
    type. A retried key returns the original response byte-identically with
...

    **Timestamps** are RFC 3339 UTC. **IDs** are
...

...
servers:
...
paths:
  /admin/api/dlq:
    get:
      tags:
      - admin
      summary: List parked jobs across environments
      operationId: adminListDeadLetters
      responses:
        '200':
          description: Parked jobs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminDeadLetter'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
  /admin/api/dlq/replay-all:
    post:
      tags:
      - admin
      summary: Replay every parked job
      operationId: adminReplayAllDeadLetters
      responses:
        '200':
          description: Replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminReplayResult'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
  /admin/api/dlq/{job_id}/replay:
    post:
      tags:
      - admin
      summary: Replay one parked job (re-enters the normal claim path)
      description: Moves the parked row back into `jobs` with a fresh attempt budget; the normal worker loop (SKIP LOCKED, per-environment fairness, delete-on-completion) runs it. Never executed inline.
      operationId: adminReplayDeadLetter
      parameters:
      - name: job_id
        in: path
        description: Job TypeID (job_…).
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminReplayResult'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No such parked job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
  /admin/api/environments:
    get:
      tags:
      - admin
      summary: List environments
      operationId: adminListEnvironments
      responses:
        '200':
          description: Environments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminEnvironment'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
    post:
      tags:
      - admin
      summary: Create an environment
      operationId: adminCreateEnvironment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateEnvironmentRequest'
        required: true
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEnvironmentDetail'
        '400':
          description: Validation error or slug already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
  /admin/api/environments/{env_id}:
    get:
      tags:
      - admin
      summary: Environment detail (includes the subscriber HMAC secret)
      operationId: adminGetEnvironment
      parameters:
      - name: env_id
        in: path
        description: Environment TypeID (env_…).
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Environment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminEnvironmentDetail'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No such environment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
  /admin/api/environments/{env_id}/api-keys:
    get:
      tags:
      - admin
      summary: List API keys for an environment (prefix only, never the key)
      operationId: adminListApiKeys
      parameters:
      - name: env_id
        in: path
        description: Environment TypeID (env_…).
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Keys.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminApiKey'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No such environment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
    post:
      tags:
      - admin
      summary: Create an API key (plaintext returned once)
      operationId: adminCreateApiKey
      parameters:
      - name: env_id
        in: path
        description: Environment TypeID (env_…).
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateApiKeyRequest'
        required: true
      responses:
        '201':
          description: Created; `key` is shown once.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiKeyCreated'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No such environment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
  /admin/api/environments/{env_id}/api-keys/{key_id}/revoke:
    post:
      tags:
      - admin
      summary: Revoke an API key (soft; row kept for audit)
      operationId: adminRevokeApiKey
      parameters:
      - name: env_id
        in: path
        description: Environment TypeID (env_…).
        required: true
        schema:
          type: string
      - name: key_id
        in: path
        description: API key TypeID (key_…).
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Revoked (now or already).
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No such API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
  /admin/api/environments/{env_id}/broadcasts:
    post:
      tags:
      - admin
      summary: Compose a broadcast (one row, fan-out on read)
      description: 'Composing is creating: the same idempotent management-plane write-path. One row per announcement, never materialized per subscriber.'
      operationId: adminCreateBroadcast
      parameters:
      - name: env_id
        in: path
        description: Environment TypeID (env_…).
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateBroadcastRequest'
        required: true
      responses:
        '200':
          description: Idempotent replay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Broadcast'
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Broadcast'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No such environment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
  /admin/api/environments/{env_id}/hmac/rotate:
    post:
      tags:
      - admin
      summary: Begin a subscriber-HMAC rotation (current → previous)
      description: |
        Generates a new current secret and moves the existing one into the
        previous slot. During the overlap BOTH secrets verify live `<Inbox />`
        sessions (auth checks current then previous), so rotation is
        zero-downtime. Complete the rotation to clear the previous slot once
        every customer backend has switched.
      operationId: adminRotateHmac
      parameters:
      - name: env_id
        in: path
        description: Environment TypeID (env_…).
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Rotated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminHmacRotation'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No such environment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
  /admin/api/environments/{env_id}/hmac/rotate/complete:
    post:
      tags:
      - admin
      summary: Complete a rotation (clear the previous secret slot)
      operationId: adminCompleteHmacRotation
      parameters:
      - name: env_id
        in: path
        description: Environment TypeID (env_…).
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Previous secret cleared.
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No such environment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
  /admin/api/environments/{env_id}/notifications:
    get:
      tags:
      - admin
      summary: Browse direct notifications (filter by subscriber, category, time)
      operationId: adminListNotifications
      parameters:
      - name: env_id
        in: path
        description: Environment TypeID (env_…).
        required: true
        schema:
          type: string
      - name: subscriber_id
        in: path
        description: Customer-provided subscriber id.
        required: true
        schema:
          type: string
          nullable: true
      - name: category
        in: path
        required: true
        schema:
          type: string
          nullable: true
      - name: after
        in: path
        description: Lower bound on `visible_at` (inclusive), RFC 3339.
        required: true
        schema:
          type: string
          format: date-time
          nullable: true
      - name: before
        in: path
        description: Upper bound on `visible_at` (exclusive), RFC 3339.
        required: true
        schema:
          type: string
          format: date-time
          nullable: true
      - name: limit
        in: path
        required: true
        schema:
          type: integer
          format: int64
          nullable: true
      - name: cursor
        in: path
        description: Opaque keyset cursor from the previous page's `next_cursor`.
        required: true
        schema:
          type: string
          nullable: true
      responses:
        '200':
          description: A page of notifications.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminNotificationPage'
        '400':
          description: Malformed cursor or out-of-range limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No such environment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
  /admin/api/environments/{env_id}/notifications/{notif_id}/timeline:
    get:
      tags:
      - admin
      summary: Status timeline for one notification (the "did it send?" answer)
      operationId: adminNotificationTimeline
      parameters:
      - name: env_id
        in: path
        description: Environment TypeID (env_…).
        required: true
        schema:
          type: string
      - name: notif_id
        in: path
        description: Notification TypeID (notif_…).
        required: true
        schema:
          type: string
      responses:
        '200':
          description: The timeline so far.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminNotificationTimeline'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No such notification.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
  /admin/api/environments/{env_id}/subscribers/{subscriber_id}:
    get:
      tags:
      - admin
      summary: 'Subscriber lookup: counters, watermarks, preferences, merged inbox'
      operationId: adminGetSubscriber
      parameters:
      - name: env_id
        in: path
        description: Environment TypeID (env_…).
        required: true
        schema:
          type: string
      - name: subscriber_id
        in: path
        description: Customer-provided subscriber id.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Subscriber view.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSubscriberView'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Authenticated but the role lacks the capability.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No such subscriber.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
  /admin/api/login:
    post:
      tags:
      - admin
      summary: Log in with email + password; starts a server-side session
      description: On success sets the `chimely_admin` session cookie (HttpOnly, SameSite=Strict, Path=/admin). Failure returns a generic error that does not reveal which field was wrong.
      operationId: adminLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
        required: true
      responses:
        '200':
          description: Logged in; sets the session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMe'
        '401':
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Missing X-Chimely-Admin header.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/api/logout:
    post:
      tags:
      - admin
      summary: 'Log out: delete the session and clear the cookie'
      operationId: adminLogout
      responses:
        '204':
          description: Logged out.
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Missing X-Chimely-Admin header.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
  /admin/api/me:
    get:
      tags:
      - admin
      summary: The signed-in admin user, role, and capabilities
      operationId: adminMe
      responses:
        '200':
          description: The current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMe'
        '401':
          description: Authentication required (no or expired session).
      security:
      - AdminSession: []
  /admin/api/users:
    get:
      tags:
      - admin
      summary: List admin users
      operationId: adminListUsers
      responses:
        '200':
          description: Users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminUserView'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Requires user:manage.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
    post:
      tags:
      - admin
      summary: Create an admin user
      operationId: adminCreateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateUserRequest'
        required: true
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserView'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required (no or expired session).
        '403':
          description: Requires user:manage.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - AdminSession: []
...
  /v1/inbox/counts:
    get:
      tags:
      - subscriber
      summary: Unread and unseen counts
      description: |
        `unread` drives list styling; `unseen` drives the bell badge (cleared
        by mark-all-seen when the inbox opens). Served from maintained counters
        (Redis cache, Postgres authoritative) — O(1), not O(rows).
      operationId: getInboxCounts
      responses:
        '200':
          description: Current counts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InboxCounts'
        '401':
          description: Missing/invalid API key or subscriber hash.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - SubscriberEnv: []
        SubscriberHash: []
        SubscriberId: []
...
components:
...
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: chimely_admin
      description: Server-side admin session cookie, set by POST /admin/api/login.
    ApiKeyBearer:
      type: http
      scheme: bearer
      description: Environment-scoped management API key.
    SubscriberEnv:
      type: apiKey
      in: header
      name: X-Chimely-Environment
    SubscriberEnvQ:
      type: apiKey
      in: query
      name: environment
    SubscriberHash:
      type: apiKey
      in: header
      name: X-Chimely-Subscriber-Hash
      description: Optional when the environment has require_subscriber_hash = false.
    SubscriberHashQ:
      type: apiKey
      in: query
      name: subscriber_hash
    SubscriberId:
      type: apiKey
      in: header
      name: X-Chimely-Subscriber
    SubscriberIdQ:
      type: apiKey
      in: query
      name: subscriber_id
...
security: []
...